feat(frontend): modularized useProfileHistoryStore
