+++ packages/frontend/src/pages/Profile/History/store/createVoteSlice.ts
@@ -1,0 +1,16 @@
+import { ActiveFilter, StateVoteSlice } from '../types'
+
+export const createVoteSlice: StateVoteSlice = (set) => ({
+    votes: {
+        activeFilter: ActiveFilter.Latest,
+        data: [],
+    },
+    setVoteActiveFilterToLatest: () =>
+        set((state) => {
+            state.votes.activeFilter = ActiveFilter.Latest
+        }),
+    setVoteActiveFilterToOldest: () =>
+        set((state) => {
+            state.votes.activeFilter = ActiveFilter.Oldest
+        }),
+})
